pricingcalc-mcp/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
boto3==1.34.0
pydantic==2.5.0
pyyaml==6.0.1
requests==2.31.0
